pipeline/requirements.txt
openai>=1.0.0
python-dotenv
httpx<0.28